Implemented custom validation error handler for pydantic models. This is done to simplyfy the client logic.
The client is now responsible for setting the invalid class on a form request, which results in a 422 error. The client loops the fieldErrors and fieldOk lists to set or remove the invalid class.

If the request is successful then it will loop all of the targets children and remove the invalid class.
All of this enables a much simpler code in python to handle form validation and user feedback as the user can identify the element with missing or wrong data.
The develop does not need to specify, which fields failed validation as the uiwiz takes care of this!