Choosing tailwind/daisyui was a mistake.
Why is that? Well to use tailwindcss/typography with daisyui it is required that we generate the tailwindcss. But it is not possible to include all classes so it is still required for us to include the precompiled daisyui/themes along with the tailwind.js file!
Because tailwind is JIT compiled only and AOT is not possible with V3 & V4...